requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
Django==5.1.4
pillow==11.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
reportlab==4.2.5
requests==2.32.3
django-bootstrap5==24.3
django-hosts==6.0
setuptools==75.8.0
openai